results/rebuttal_new_baselines_llmatic_missing/run-20231111-034044_LLMatic-zero_shot_donnemartin-system-design-oop-recipe_0_20-runs_log/recipe/13/LLMatic/recipe.py
class Recipe:
	def __init__(self, name, ingredients, instructions, images, categories):
		self.name = name
		self.ingredients = ingredients
		self.instructions = instructions
		self.images = images
		self.categories = categories

	def edit_recipe(self, name, ingredients, instructions, images, categories):
		self.name = name
		self.ingredients = ingredients
		self.instructions = instructions
		self.images = images
		self.categories = categories

	def delete_recipe(self):
		self.name = None
		self.ingredients = None
		self.instructions = None
		self.images = None
		self.categories = None


class RecipeValidator:
	@staticmethod
	def validate(recipe):
		if not recipe.name or not recipe.ingredients or not recipe.instructions:
			return False
		return True
